refactor: simplify genesis configuration and file loading
- Add `Default` derive for `GenesisFileConfig`
- Simplify genesis file loading in `StratusStorage`
- Remove unused `find_genesis_file` method from `GenesisConfig`
- Add `from_str_hex` method to `Wei` for robust hex parsing
- Clean up unused imports and code structure
